+++ .jabbarc
@@ -1,0 +1,1 @@
+"openjdk@1.11.0-2"
