+++ src/main/java/Helpers/NoArgsCallbackFunction.java
