+++ src/main/java/Maze/Candy/Candy.java
