+++ src/main/java/Maze/Candy/CandyPowerType.java
