+++ src/main/java/Maze/Candy/CandyTest.java
