+++ src/main/java/Maze/Candy/PointsCandy.java
