+++ src/main/java/Maze/Candy/TimeCandy.java
