+++ src/main/java/Maze/ELocation.java
