+++ src/main/java/Maze/ELocationCell.java
