+++ src/main/java/Maze/ELocationType.java
