+++ src/main/java/Maze/MazeBuilder/Exceptions/MazeBuilderException.java
@@ -1,0 +1,12 @@
+package Maze.MazeBuilder.Exceptions;
+
+import Maze.MazeBuilder.RectangleMazeBuilder;
+
+public class MazeBuilderException extends Throwable {
+    private final RectangleMazeBuilder rectangleMazeBuilder;
+
+    public MazeBuilderException(RectangleMazeBuilder rectangleMazeBuilder, String message) {
+        super(message);
+        this.rectangleMazeBuilder = rectangleMazeBuilder;
+    }
+}
