+++ src/main/java/Maze/MazeGenerator/Exceptions/CandyBuilderException.java
