+++ src/main/java/Maze/MazeGenerator/Exceptions/PortalCandyBuilderException.java
