+++ src/main/java/Maze/MazeGenerator/GenerateCandyConfig.java
