+++ src/main/java/Maze/Solver/Adapter/SolverAdapter.java
