+++ src/main/java/Maze/Solver/BFS/BFSCell.java
