+++ src/main/java/Maze/Solver/BFS/BFSSolverAdapter.java
