+++ src/main/java/Maze/Solver/DFS/DFSSolverAdapter.java
