+++ src/main/java/Maze/Solver/DFS/PathNeighbourResult.java
