+++ src/main/java/UI/MazePreviewPanel.java
