+++ src/main/java/player/BasePlayer.java
